fix: el resguardo de stock ahora funciona para productos por peso
El fix de 59b972b8 comparaba con parseInt. Como `stock` es numeric y los
productos con allowDecimal viven en valores fraccionados (un pedido de 0.5 kg
los deja en 9.5), parseInt("9.5") = 9 nunca coincidia con 9.5. Resultado: para
TODO producto por peso el resguardo daba siempre "el usuario toco el stock",
asi que no los protegia de nada, y encima mandaba el 9 truncado. Editarle el
precio a un producto por peso le comia media unidad.

Va parseFloat en el chequeo y en el valor que se manda. El fallback `|| 0`
sigue cubriendo el string vacio (parseFloat("") es NaN igual que parseInt).
La validacion previa tambien pasa a parseFloat: con parseInt, "-0.4" se leia
como -0, pasaba el chequeo de negativos y ahora se guardaria en negativo.

+++ src/components/ProductManagement.tsx
@@ -241,7 +241,9 @@ export function ProductManagement({ accessToken, onBack, onManageCategories, onM
       toast.error('El precio debe ser mayor o igual a 0');
       return;
     }
-    if (!formData.unlimitedStock && (!formData.stock || parseInt(formData.stock) < 0)) {
+    // parseFloat acá también: con parseInt, "-0.4" se leía como -0 y pasaba la
+    // validación, y ahora que el valor se guarda con decimales entraría negativo.
+    if (!formData.unlimitedStock && (!formData.stock || parseFloat(formData.stock) < 0)) {
       toast.error('El stock debe ser mayor o igual a 0');
       return;
     }
@@ -261,9 +263,14 @@ export function ProductManagement({ accessToken, onBack, onManageCategories, onM
       const eraIlimitadoAntes = editingProduct
         ? (editingProduct.unlimitedStock === true || editingProduct.stock === -1)
         : false;
+      // parseFloat y no parseInt: `stock` es numeric en Postgres y un producto
+      // con allowDecimal se queda en valores fraccionados (los pedidos le restan
+      // 0.5). Con parseInt, parseInt("9.5") = 9 nunca coincidía con 9.5, así que
+      // para TODO producto por peso este chequeo daba "se tocó" y además mandaba
+      // el 9 truncado: editar solo el precio le comía media unidad.
       const stockSeToco = !editingProduct
         || formData.unlimitedStock !== eraIlimitadoAntes
-        || parseInt(formData.stock) !== editingProduct.stock;
+        || parseFloat(formData.stock) !== editingProduct.stock;
 
       const productData = {
         name: formData.name.trim(),
@@ -291,7 +298,7 @@ export function ProductManagement({ accessToken, onBack, onManageCategories, onM
           };
         }),
         ...(stockSeToco
-          ? { stock: formData.unlimitedStock ? 0 : (parseInt(formData.stock) || 0) }
+          ? { stock: formData.unlimitedStock ? 0 : (parseFloat(formData.stock) || 0) }
           : {})
       };
